add explicit cast

+++ collector/kernel/bpf_src/tcp-processor/bpf_types.h
@@ -74,9 +74,9 @@ typedef u64 TIMESTAMP;
     /* Check if key already exists to distinguish duplicate vs table full */                                                   \
     void *existing = bpf_map_lookup_elem(&SAVED_ARGS_TABLE(FUNC), &SAVED_ARGS_TABLE_KEY);                                      \
     if (existing) {                                                                                                            \
-      bpf_trace_printk(#FUNC ": duplicate arg insert. pid_tgid=%llx", bpf_get_current_pid_tgid());                                                                        \
+      bpf_trace_printk(#FUNC ": duplicate arg insert. pid_tgid=%llx", (unsigned long long)(bpf_get_current_pid_tgid()));        \
     } else {                                                                                                                   \
-      bpf_trace_printk(#FUNC ": args table is full, dropped insert. pid_tgid=%llx", bpf_get_current_pid_tgid());                                                          \
+      bpf_trace_printk(#FUNC ": args table is full, dropped insert. pid_tgid=%llx", (unsigned long long)(bpf_get_current_pid_tgid()));                                                          \
     }                                                                                                                          \
   }                                                                                                                            \
   }
